requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-mistralai>=0.1.0
mistralai>=1.0.0
pypdf2>=3.0.0
python-pptx>=0.6.21
pillow>=10.0.0
pytesseract>=0.3.10
networkx>=3.2.0
plotly>=5.17.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
sentence-transformers>=2.2.0
graphviz>=0.20.0





